pyproject.toml
[project]
name = "rag-based-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "langchain",
    "langchain-community",
    "langchain-groq",
    "sentence-transformers",
    "pydantic",
    "pydantic-core",
    "pinecone",
    "pypdf",
    "python-dotenv",
    "rank-bm25",
]
